Extract shared status tag presenter to Bops Core
