+++ engines/bops_consultees/app/views/bops_consultees/planning_applications/show.html.erb
@@ -10,6 +10,7 @@
 
     <p class="govuk-body">
       <%= t(".application_number") %> <strong><%= @planning_application.reference %></strong>
+      <%= @planning_application.status_tag %>
     </p>
   </div>
 </div>
